My VPN/bot/requirements.txt
python-telegram-bot[job-queue]==21.7
qrcode[pil]==7.4.2
